Benchmark expected-value LCG overflows a 32-bit long (#27)
* Compute the benchmark LCG in unsigned arithmetic

The expected-value LCG in coucal_test() overflowed a signed 32-bit long on
ILP32, which UBSan flags as soon as the abi32 leg is built with sanitizers.
Wrap in unsigned instead; LP64 values are unchanged.

Closes #26


* Apply clang-format to the changed line


---------

+++ tests.c
@@ -146,7 +146,9 @@ static int coucal_test(const char *snum) {
         int result = 0;
         char buffer[256];
         const char *name;
-        const long expected = (long) i * 1664525 + 1013904223;
+        /* unsigned: this LCG overflows a 32-bit long on ILP32 */
+        const long expected =
+            (long) ((unsigned long) i * 1664525UL + 1013904223UL);
         if (strings == NULL) {
           snprintf(buffer, sizeof(buffer),
             "http://www.example.com/website/sample/for/hashtable/"
